requirements.txt
aiosqlite==0.22.1
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
betterproto==2.0.0b7
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
diff_cover==9.6.0
ffmpy==1.0.0
fsspec==2026.1.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
macholib==1.16.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mashumaro==3.17
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.3
protobuf3-to-dict==0.1.5
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pygame==2.6.1
Pygments==2.19.2
pyiceberg==0.10.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.11
PyJWT==2.10.1
pymunk==7.2.0
pyparsing==3.3.2
pyroaring==1.0.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-socks==2.8.0
PyYAML==6.0.2
realtime==2.27.2
regex==2025.7.34
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlfluff==3.4.2
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tblib==3.1.0
tenacity==9.1.2
TikTokLive==6.6.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
websockets_proxy==0.1.3
wheel==0.46.2
yarl==1.22.0
